Moving VideList into List folder
